seeder/requirements.txt
redis
mysql-connector-python
